+++ src/main/kotlin/org/wfanet/measurement/common/db/r2dbc/BUILD.bazel
@@ -8,7 +8,6 @@ kt_jvm_library(
     deps = [
         "//imports/java/com/google/protobuf",
         "//imports/java/io/r2dbc",
-        "//imports/java/org/postgresql:r2dbc",
         "//imports/java/org/reactivestreams",
         "//imports/kotlin/com/google/protobuf/kotlin",
         "//imports/kotlin/kotlinx/coroutines:core",
